requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
xlrd>=2.0.1
anthropic>=0.80.0